chore: Seperate crud functions (#1843)
* chore: separate crud functions for more readability

* fix: format
