+++ app/db/crud/admin.py
@@ -1,0 +1,203 @@
+from datetime import datetime, timezone
+
+from sqlalchemy import select
+from sqlalchemy.ext.asyncio import AsyncSession
+
+from app.db.models import Admin, AdminUsageLogs
+from app.models.admin import AdminCreate, AdminModify
+
+
+async def load_admin_attrs(admin: Admin):
+    await admin.awaitable_attrs.users
+    await admin.awaitable_attrs.usage_logs
+
+
+async def get_admin(db: AsyncSession, username: str) -> Admin:
+    """
+    Retrieves an admin by username.
+
+    Args:
+        db (AsyncSession): Database session.
+        username (str): The username of the admin.
+
+    Returns:
+        Admin: The admin object.
+    """
+    admin = (await db.execute(select(Admin).where(Admin.username == username))).unique().scalar_one_or_none()
+    if admin:
+        await load_admin_attrs(admin)
+    return admin
+
+
+async def create_admin(db: AsyncSession, admin: AdminCreate) -> Admin:
+    """
+    Creates a new admin in the database.
+
+    Args:
+        db (AsyncSession): Database session.
+        admin (AdminCreate): The admin creation data.
+
+    Returns:
+        Admin: The created admin object.
+    """
+    db_admin = Admin(**admin.model_dump(exclude={"password"}), hashed_password=admin.hashed_password)
+    db.add(db_admin)
+    await db.commit()
+    await db.refresh(db_admin)
+    await load_admin_attrs(db_admin)
+    return db_admin
+
+
+async def update_admin(db: AsyncSession, db_admin: Admin, modified_admin: AdminModify) -> Admin:
+    """
+    Updates an admin's details.
+
+    Args:
+        db (AsyncSession): Database session.
+        dbadmin (Admin): The admin object to be updated.
+        modified_admin (AdminModify): The modified admin data.
+
+    Returns:
+        Admin: The updated admin object.
+    """
+    if modified_admin.is_sudo is not None:
+        db_admin.is_sudo = modified_admin.is_sudo
+    if modified_admin.is_disabled is not None:
+        db_admin.is_disabled = modified_admin.is_disabled
+    if modified_admin.hashed_password is not None and db_admin.hashed_password != modified_admin.hashed_password:
+        db_admin.hashed_password = modified_admin.hashed_password
+        db_admin.password_reset_at = datetime.now(timezone.utc)
+    if modified_admin.telegram_id is not None:
+        db_admin.telegram_id = modified_admin.telegram_id
+    if modified_admin.discord_webhook is not None:
+        db_admin.discord_webhook = modified_admin.discord_webhook
+    if modified_admin.discord_id is not None:
+        db_admin.discord_id = modified_admin.discord_id
+    if modified_admin.sub_template is not None:
+        db_admin.sub_template = modified_admin.sub_template
+    if modified_admin.sub_domain is not None:
+        db_admin.sub_domain = modified_admin.sub_domain
+    if modified_admin.support_url is not None:
+        db_admin.support_url = modified_admin.support_url
+    if modified_admin.profile_title is not None:
+        db_admin.profile_title = modified_admin.profile_title
+
+    await db.commit()
+    await load_admin_attrs(db_admin)
+    return db_admin
+
+
+async def remove_admin(db: AsyncSession, dbadmin: Admin) -> None:
+    """
+    Removes an admin from the database.
+
+    Args:
+        db (AsyncSession): Database session.
+        dbadmin (Admin): The admin object to be removed.
+    """
+    await db.delete(dbadmin)
+    await db.commit()
+
+
+async def get_admin_by_id(db: AsyncSession, id: int) -> Admin:
+    """
+    Retrieves an admin by their ID.
+
+    Args:
+        db (AsyncSession): Database session.
+        id (int): The ID of the admin.
+
+    Returns:
+        Admin: The admin object.
+    """
+    admin = (await db.execute(select(Admin).where(Admin.id == id))).unique().scalar_one_or_none()
+    if admin:
+        await load_admin_attrs(admin)
+    return admin
+
+
+async def get_admin_by_telegram_id(db: AsyncSession, telegram_id: int) -> Admin:
+    """
+    Retrieves an admin by their Telegram ID.
+
+    Args:
+        db (AsyncSession): Database session.
+        telegram_id (int): The Telegram ID of the admin.
+
+    Returns:
+        Admin: The admin object.
+    """
+    admin = (await db.execute(select(Admin).where(Admin.telegram_id == telegram_id))).unique().scalar_one_or_none()
+    if admin:
+        await load_admin_attrs(admin)
+    return admin
+
+
+async def get_admin_by_discord_id(db: AsyncSession, discord_id: int) -> Admin:
+    """
+    Retrieves an admin by their Discord ID.
+
+    Args:
+        db (AsyncSession): Database session.
+        discord_id (int): The Discord ID of the admin.
+
+    Returns:
+        Admin: The admin object.
+    """
+    admin = (await db.execute(select(Admin).where(Admin.discord_id == discord_id))).unique().scalar_one_or_none()
+    if admin:
+        await load_admin_attrs(admin)
+    return admin
+
+
+async def get_admins(
+    db: AsyncSession, offset: int | None = None, limit: int | None = None, username: str | None = None
+) -> list[Admin]:
+    """
+    Retrieves a list of admins with optional filters and pagination.
+
+    Args:
+        db (AsyncSession): Database session.
+        offset (Optional[int]): The number of records to skip (for pagination).
+        limit (Optional[int]): The maximum number of records to return.
+        username (Optional[str]): The username to filter by.
+
+    Returns:
+        List[Admin]: A list of admin objects.
+    """
+    query = select(Admin)
+    if username:
+        query = query.where(Admin.username.ilike(f"%{username}%"))
+    if offset:
+        query = query.offset(offset)
+    if limit:
+        query = query.limit(limit)
+
+    admins = (await db.execute(query)).scalars().all()
+
+    for admin in admins:
+        await load_admin_attrs(admin)
+
+    return admins
+
+
+async def reset_admin_usage(db: AsyncSession, db_admin: Admin) -> Admin:
+    """
+    Retrieves an admin's usage by their username.
+    Args:
+        db (AsyncSession): Database session.
+        db_admin (Admin): The admin object to be updated.
+    Returns:
+        Admin: The updated admin.
+    """
+    if db_admin.used_traffic == 0:
+        return db_admin
+
+    usage_log = AdminUsageLogs(admin=db_admin, used_traffic_at_reset=db_admin.used_traffic)
+    db.add(usage_log)
+    db_admin.used_traffic = 0
+
+    await db.commit()
+    await db.refresh(db_admin)
+    await load_admin_attrs(db_admin)
+    return db_admin
