+++ app/operation/host.py
@@ -5,7 +5,7 @@ from app.db.models import ProxyHost
 from app.models.host import CreateHost, BaseHost
 from app.models.admin import AdminDetails
 from app.operation import BaseOperation
-from app.db.crud import create_host, get_host_by_id, remove_host, get_hosts, modify_host
+from app.db.crud.host import create_host, get_host_by_id, remove_host, get_hosts, modify_host
 from app.core.hosts import hosts as hosts_storage
 from app.utils.logger import get_logger
