+++ app/telegram/middlewares/acl.py
@@ -3,7 +3,8 @@ from typing import Any, Awaitable, Callable, Dict
 from aiogram import BaseMiddleware
 from aiogram.types import Update
 
-from app.db import GetDB, crud
+from app.db import GetDB
+from app.db.crud.admin import get_admin_by_telegram_id
 from app.models.admin import AdminDetails
 
 
@@ -14,7 +15,7 @@ class ACLMiddleware(BaseMiddleware):
         message_obj = event.message or event.callback_query or event.inline_query
         user_id = message_obj.from_user.id
         async with GetDB() as db:
-            admin = await crud.get_admin_by_telegram_id(db, user_id)
+            admin = await get_admin_by_telegram_id(db, user_id)
             if admin:
                 if admin.is_disabled:
                     await message_obj.reply("your account hase been disabled.")
