+++ app/utils/jwt.py
@@ -7,7 +7,7 @@ from math import ceil
 
 from aiocache import cached
 from app.db import GetDB
-from app.db.crud import get_jwt_secret_key
+from app.db.crud.general import get_jwt_secret_key
 from config import JWT_ACCESS_TOKEN_EXPIRE_MINUTES
